requirements.txt
mediapipe
opencv-python
numpy
pandas
scikit-learn
pygame
gTTS
pydobot
pyserial